third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/events/messageevent-constructor.https-expected.txt

This is a testharness.js-based test.
PASS Default event values
PASS MessageEventInit dictionary
PASS Passing null for ports member
PASS ports attribute should be a FrozenArray
PASS initMessageEvent operation
FAIL Passing null for ports parameter to initMessageEvent assert_throws: function "function() {
    ev.initMessageEvent("test", true, false, "testData", "testOrigin", "testId", window, null)
  }" did not throw
FAIL initMessageEvent operation default parameter values assert_equals: data attribute expected (object) null but got (undefined) undefined
PASS Passing ServiceWorker for source member
Harness: the test ran to completion.
